// src/app/core/interfaces/icredit.ts
export interface Cast {
	adult: boolean;
	gender: number;
	id: number;
	known_for_department: string;
	name: string;
	original_name: string;
	popularity: number;
	profile_path: string;
	character: string;
	credit_id: string;
	order: number;
}

export interface Crew {
	adult: boolean;
	gender: number;
	id: number;
	known_for_department: string;
	name: string;
	original_name: string;
	popularity: number;
	profile_path?: any;
	credit_id: string;
	department: string;
	job: string;
}

export interface Icredit {
	cast: Cast[];
	crew: Crew[];
	id: number;
}