// src/app/core/interfaces/ishow.ts
export interface IShow {

    adult: boolean;
	backdrop_path: string;
	genre_ids: number[];
	id: number;
	original_language: string;
	original_title: string;
	overview: string;
	popularity: number;
	poster_path: string;
	release_date: string;
	title: string;
	video: boolean;
	vote_average: number;
	vote_count: number;


	//  tv shows
  name: string
  original_name: string
//   poster_path: string
  media_type: string
  first_air_date: string
  origin_country: string[]




 
  credit_id: string;
  department: string;
  job: string;




}